Move the MechanicalMatrix class from MultiMatrix.h to MechanicalMatrix.h
