+++ SofaKernel/modules/SofaCore/src/sofa/core/behavior/MultiMatrix.h
@@ -19,51 +19,13 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#ifndef SOFA_CORE_BEHAVIOR_MULTIMATRIX_H
-#define SOFA_CORE_BEHAVIOR_MULTIMATRIX_H
+#pragma once
 
 #include <sofa/core/MultiVecId.h>
+#include <sofa/core/behavior/MechanicalMatrix.h>
 
-namespace sofa
+namespace sofa::core::behavior
 {
-
-namespace core
-{
-
-namespace behavior
-{
-
-/// Helper class allowing to construct mechanical expressions
-///
-class SOFA_CORE_API MechanicalMatrix
-{
-protected:
-    enum { MFACT = 0, BFACT = 1, KFACT = 2 };
-    defaulttype::Vec<3,SReal> factors;
-public:
-    MechanicalMatrix(SReal m, SReal b, SReal k) : factors(m,b,k) {}
-    explicit MechanicalMatrix(const defaulttype::Vec<3,SReal>& f) : factors(f) {}
-
-    static const MechanicalMatrix M;
-    static const MechanicalMatrix B;
-    static const MechanicalMatrix K;
-
-    SReal getMFact() const { return factors[MFACT]; }
-    SReal getBFact() const { return factors[BFACT]; }
-    SReal getKFact() const { return factors[KFACT]; }
-
-    MechanicalMatrix operator + (const MechanicalMatrix& m2) const { return MechanicalMatrix(factors + m2.factors); }
-    MechanicalMatrix operator - (const MechanicalMatrix& m2) const { return MechanicalMatrix(factors - m2.factors); }
-    MechanicalMatrix operator - () const { return MechanicalMatrix(- factors); }
-    MechanicalMatrix operator * (SReal f) const { return MechanicalMatrix(factors * f); }
-    MechanicalMatrix operator / (SReal f) const { return MechanicalMatrix(factors / f); }
-
-    /// Declares that the operator<< is our friend so he can access the our private/protected members
-    friend SOFA_CORE_API std::ostream& operator << (std::ostream& out, const MechanicalMatrix& m );
-};
-
-/// Declares that there is an function operator<< for MechanicalMatrix.
-SOFA_CORE_API std::ostream& operator << (std::ostream& out, const MechanicalMatrix& m );
 
 /// Helper class providing a high-level view of underlying linear system matrices.
 ///
@@ -123,10 +85,4 @@ public:
     }
 };
 
-} // namespace behavior
-
-} // namespace core
-
-} // namespace sofa
-
-#endif
+} /// namespace sofa::core::behavior
